# Optimize octree and fix leak