--- a/servers/visual_server.cpp
+++ b/servers/visual_server.cpp
@@ -2432,6 +2432,10 @@ VisualServer::VisualServer() {
 
 	GLOBAL_DEF("rendering/quality/filters/use_nearest_mipmap_filter", false);
 
+	const char *sz_balance_render_tree = "rendering/quality/spatial_partitioning/render_tree_balance";
+	GLOBAL_DEF(sz_balance_render_tree, 0.17f);
+	ProjectSettings::get_singleton()->set_custom_property_info(sz_balance_render_tree, PropertyInfo(Variant::REAL, sz_balance_render_tree, PROPERTY_HINT_RANGE, "0.0,1.0,0.01"));
+
 	GLOBAL_DEF("rendering/batching/options/use_batching", true);
 	GLOBAL_DEF_RST("rendering/batching/options/use_batching_in_editor", true);
 	GLOBAL_DEF("rendering/batching/options/single_rect_fallback", false);