… core/math/octree.h
… core/math/octree_definition.inc