## Octant Limit
Prevents adding new octants until a limiting number of elements have been added to the current octant. This enables balancing the benefits of brute force against the benefits of spatial partitioning. The limit can be set per octree.

Project settings are added for the rendering octree only to set the best balance per project depending on number of tests per frame / tick, and the amount of editing of the octree.

## Octant Leak
Fixes octants being leaked when removing elements.
_This is mirroring @naithar fix for the leak #41122, so it can be self contained.
On suggestion from @clayjohn I have updated the leak fix to be similar to the other PR, and added naithar as co-author in case this ends up getting merged together, to make things easier for Akien._

## Cached linear local_vectors
(second commit) Instead of using linked lists at runtime this maintains a dirty flag per octant and builds on demand a linear list of Element pointers and AABBs which can be used for quicker cull tests. This can lead to significantly better test performance.

This is offered as a second version of the octree, Octree_CL. It is only used for the rendering octree currently, as it has only been shown to offer performance increases there so far.

# Background
As discussed with reduz, this prevents adding new octants until a limiting number of elements have been added to the current octant, and is probably a better option than directly limiting the depth of the tree. This enables balancing the benefits of brute force against the benefits of spatial partitioning. The limit can be set per octree.

Preliminary tests indicate that for visibility culling brute force is significantly faster than using the octree as is.

The implementation so far is very simple, it doesn't split octants until the limit is reached. This means that on splitting, some of the elements in the parent octant may have been better off in the children. I have written code to rebalance the tree on splitting, but it isn't significantly faster so is left out at this stage.

Scenes which particularly benefit are those containing large numbers of objects, especially large number of moving objects / created / erased. It can also significantly speed up cull tests with the cached lists. Most 3d games will see a modest improvement, but some may gain a 2x improvement in fps or more. This is highly dependent on to what degree octree is a bottleneck in particular game.

Fixes #38142
Fixes #41480
Related to #40059

## Notes
* The PR is less extensive than it looks in terms of line numbers, because it moves most of the existing octree into a separate file for multiple inclusion, and the addition of local_vector from Godot 4.0.
* This is made against 3.2, there will be a bit of porting to get it to work with master because of changes to geometry. Also there may be further modifications to the version in the 4.x branch in future as we move to other spatial partitioning schemes.
* It also adds some basic debugging facilities to the octree (TOOLS only).
* I've experimentally tried a system to automatically set the limit depending on the use case - but it is a very difficult problem. For instance some frames may do more tests than others, or a lot of tests on e.g. level startup. I've tried maintaining a running average but the more I think about it the more I tend to think we should have this as a project setting, set to a reasonable value, and then tweakable for better performance. I'm aware of the project settings bloat problem but I can't see an easy better way around this, so I'm going to add this to the PR.

### Local Vector
This PR introduces `local_vector.h` from Godot 4.x. This might be better as a separate commit so the octree changes could be reverted without the local_vector being reverted, if something else is relying on local_vector. Note that PouleyKetchoupp's PR #40313 also contains local_vector.

We could potentially merge local_vector as a separate commit first, then me and Pouley could modify our PRs. Leave this up to Akien though as to whatever works best.

## Demo
To try it out, load this project, and try changing the project settings `rendering/quality/spatial_partitioning/render_tree_balance`.

* At a value of 0.0 I get 26fps, and at 1.0 I get 240fps.
* For reference, with the old octree I get somewhere around 23fps.
* The relative speedup in this example depends how many can be culled, when all are culled the speedup is higher, when all are showing, less so, because the bottlenecks are elsewhere.
* In most cases for rendering, shifting the balance towards 1.0 (in favour of octree editing rather than tests) will give best performance. Tests only start to become the bottleneck when you are doing lots of them (100s or 1000s per frame), in which case shifting towards 0.0 can help (and the benefits of the linear cache start to really show).
* A balance of 0.0 is closest to the old octree behaviour. However it should still be faster due to linear caching which is always on.

[octree_test.zip](https://github.com/godotengine/godot/files/5052230/octree_test.zip)