@clayjohn 2020-08-12
I tested out this PR a little bit this evening. Unfortunately, I spotted a regression. 

In the TPS demo, just moving the character at full speed causes the character to disappear until they are moved back to the original position, and then if you walk slowly they will not disappear. I am guess something is getting lost when objects move too fast.

The performance was slightly better with ``world/3d/spatial_partitioning/render_tree_balance`` set to 1.0. Although, I didn't expect much of an improvement as there are not very many moving objects. And not much is being created or destroyed.