@Calinou 2020-08-12
If anyone needs the tps-demo, you can download it from here (link valid until September 2020): https://0x0.st/iY8e.zip