@clayjohn 2020-08-12
I tested the TPS demo on the latest commit and it appears to work great! :)