@akien-mga 2020-09-30
Alright, time to merge and get this battle tested in 3.2.4 beta 1 :)

@reduz says the changes seem fine but he's not so familiar with this code anymore to give an in-depth review, yet he trusts the extensive work done by @lawnjelly and @naithar.

So let's merge and see what happens, we should have plenty of time until 3.2.4-stable to catch potential regressions, if any.