@jamie-pate 2020-10-03
I was hopeful this would fix my bottleneck, but I think it's related to 'pairing' which you mentioned is left out.
![image](https://user-images.githubusercontent.com/682658/94998992-ec0ed780-056a-11eb-8cfe-c31f0a679a4d.png)
Unfortunately I don't get what 'pairing' means in this context or why it's running so slow :)

![image](https://user-images.githubusercontent.com/682658/94999008-0648b580-056b-11eb-9883-a1b65826842f.png)