@lawnjelly 2020-10-04
If you create an issue with a minimum reproduction project, it is more likely to get looked at (I might be able to profile it and find something obvious for example).