@jamie-pate 2020-10-04
Done

I'm be curious at what your method is for profiling? I've only hacked in some statistics to the object visibility calls. I guess with it toggling every frame any old profiling will do? (I was trying to catch stuff that only happens every once in a while, which is harder)
[object-profiling.diff.txt](https://github.com/godotengine/godot/files/5323763/object-profiling.diff.txt)